Refactor PopulateValidationParametersWithCurrentConfigurationAsync to a shared utility class
